# open-api-spec/openapi.json
{
  "openapi": "3.0.2",
  "info": {
    "title": "Anubis",
    "description": "Anubis is a flexible Policy Enforcement solution that makes easier to reuse security policies across different services, assuming the policies entail the same resource.",
    "version": "0.5.0-dev"
  },
  "paths": {
...
    "/v1/policies/access-modes": {
      "get": {
        "tags": [
          "policies"
        ],
        "summary": "List supported Access Modes",
        "operationId": "read_modes_v1_policies_access_modes_get",
        "parameters": [
          {
            "required": false,
            "schema": {
              "title": "Skip",
              "type": "integer",
              "default": 0
            },
            "name": "skip",
            "in": "query"
          },
          {
            "required": false,
            "schema": {
              "title": "Limit",
              "type": "integer",
              "default": 100
            },
            "name": "limit",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Read Modes V1 Policies Access Modes Get",
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Mode"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/policies/agent-types": {
      "get": {
        "tags": [
          "policies"
        ],
        "summary": "List supported Agent Types",
        "operationId": "read_modes_v1_policies_agent_types_get",
        "parameters": [
          {
            "required": false,
            "schema": {
              "title": "Skip",
              "type": "integer",
              "default": 0
            },
            "name": "skip",
            "in": "query"
          },
          {
            "required": false,
            "schema": {
              "title": "Limit",
              "type": "integer",
              "default": 100
            },
            "name": "limit",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Read Modes V1 Policies Agent Types Get",
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AgentType"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AgentType": {
        "title": "AgentType",
        "required": [
          "iri",
          "name"
        ],
        "type": "object",
        "properties": {
          "iri": {
            "title": "Iri",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        }
      },
...
      "HTTPValidationError": {
        "title": "HTTPValidationError",
        "type": "object",
        "properties": {
          "detail": {
            "title": "Detail",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            }
          }
        }
      },
      "Mode": {
        "title": "Mode",
        "required": [
          "iri",
          "name"
        ],
        "type": "object",
        "properties": {
          "iri": {
            "title": "Iri",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        }
      },
...
      "ValidationError": {
        "title": "ValidationError",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "type": "object",
        "properties": {
          "loc": {
            "title": "Location",
            "type": "array",
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            }
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        }
      }
    },
    "securitySchemes": {
      "OptionalHTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}